backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart